Bugfix: fix segfault in pelogue plugin. Bugfix: fix no allocation for step problem (#238). Bugfix: fix invalid local node id leading to failed startup (#240). Bugfix: don't execute expilogue for every step (#239). Enhancement: isolate service processes from user environment (#233). Enhancement:  use uint32_t for I/O rank handling. Enhancement: report the correct count of gres resources even if some are     missing (#232). Update regression test suite. Get rid of obsolete psslurmaccount.c file.

+++ plugins/psslurm/src/psslurm/psslurmcomm.c
@@ -1170,12 +1170,13 @@ void srunEnableIO(Step_t *step)
     enabled = 1;
 }
 
-int srunSendIO(uint16_t type, uint16_t taskid, Step_t *step,
+int srunSendIO(uint16_t type, uint32_t taskid, Step_t *step,
                 char *buf, uint32_t bufLen)
 {
     int ret, error;
 
-    ret = srunSendIOSock(type, taskid, step->srunIOMsg.sock, buf, bufLen, &error);
+    ret = srunSendIOSock(type, taskid, step->srunIOMsg.sock, buf,
+			    bufLen, &error);
 
     if (ret < 0 ) {
 	switch (error) {
@@ -1191,7 +1192,7 @@ int srunSendIO(uint16_t type, uint16_t taskid, Step_t *step,
     return ret;
 }
 
-int srunSendIOSock(uint16_t type, uint16_t taskid, int sock,
+int srunSendIOSock(uint16_t type, uint32_t taskid, int sock,
                 char *buf, uint32_t bufLen, int *error)
 {
     PS_DataBuffer_t data = { .buf = NULL };
